LIBFCREPO-1189. Moved RDF property validation into its own module.
https://umd-dit.atlassian.net/browse/LIBFCREPO-1189
